+++ src/main/java/com/java8/test/chapter01/FilterApplesTest.java
@@ -1,13 +1,13 @@
 package com.java8.test.chapter01;
 
-import com.java8.dto.Apple;
+import com.java8.model.Apple;
 import com.java8.enu.COLOR;
 
 import java.util.ArrayList;
 import java.util.List;
 
 public class FilterApplesTest {
-    private static final Double MIN_HEAVY_WEIGHT = 150.0;
+    private static final int MIN_HEAVY_WEIGHT = 150;
 
     public static boolean isGreenApple(Apple apple) {
         return COLOR.GREEN.equals(apple.getColor());
@@ -35,7 +35,7 @@ public class FilterApplesTest {
 
     public static void main(String[] args) {
         List<Apple> appleList = new ArrayList<>();
-        Apple a = new Apple("A", COLOR.GREEN, 100.0);
+        Apple a = new Apple("A", COLOR.GREEN, 100);
         appleList.add(a);
         Apple b = new Apple("B", COLOR.RED, MIN_HEAVY_WEIGHT+1);
         appleList.add(b);
