+++ src/main/java/com/java8/test/chapter01/FilterApplesTest02.java
@@ -1,6 +1,6 @@
 package com.java8.test.chapter01;
 
-import com.java8.dto.Apple;
+import com.java8.model.Apple;
 import com.java8.enu.COLOR;
 
 import java.util.ArrayList;
@@ -10,11 +10,11 @@ import static java.util.Comparator.comparing;
 import static java.util.stream.Collectors.toList;
 
 public class FilterApplesTest02 {
-    private static final Double MIN_HEAVY_WEIGHT = 150.0;
+    private static final int MIN_HEAVY_WEIGHT = 150;
 
     public static void main(String[] args) {
         List<Apple> appleList = new ArrayList<>();
-        Apple a = new Apple("A", COLOR.GREEN, 200.0);
+        Apple a = new Apple("A", COLOR.GREEN, 200);
         appleList.add(a);
         Apple b = new Apple("B", COLOR.RED, MIN_HEAVY_WEIGHT+1);
         appleList.add(b);
